requirements.txt
pydantic>=2.0.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pytest>=8.0.0

# Opcionales: backends reales de sentiment
# azure-ai-textanalytics>=5.3.0
# anthropic>=0.40.0
# openai>=1.50.0
